+++ docs/changelog.md
@@ -11,6 +11,10 @@
 
 ## 更新说明
 
+1.4.90 add: add macro_china_bank_financing interface
+
+    1. 新增 macro_china_bank_financing 接口，获取银行理财产品发行数量数据
+
 1.4.89 fix: fix stock_us_hist interface
 
     1. 修复 stock_us_hist 接口，结果数据根据日期排序
@@ -212,6 +216,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.4.90 add: add macro_china_bank_financing interface
 
 1.4.89 fix: fix stock_us_hist interface